--- python/tokenspeed/runtime/disaggregation/base/poll.py
@@ -1,40 +1,0 @@
-# Copyright (c) 2026 LightSeek Foundation
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from __future__ import annotations
-
-
-class TransferPoll:
-    """Transfer-status FSM shared by both disaggregation roles.
-
-    Integer-ordered so :meth:`DisaggManagerBase.update_status` can enforce
-    monotonic, ``Failed``-sticky transitions via ``max()``. The KV
-    (prefill->decode) and embedding (encode->prefill) paths use this single
-    enum; ``WaitingForInput`` is only meaningful to the KV path (the embedding
-    path never sets it, but sharing one ordered enum keeps the two status FSMs
-    from drifting apart).
-    """
-
-    Failed = 0
-    Bootstrapping = 1
-    Bootstrapped = 2
-    WaitingForInput = 3
-    Transferring = 4
-    Success = 5
