+++ python/tokenspeed/runtime/pd/epd/embedding_transfer.py
@@ -26,17 +26,15 @@ per item over the same Mooncake RDMA engine the KV path uses. The prefill side
 is the receiver that pre-registers buffers and registers with the bootstrap
 server (direction reversed relative to prefill->decode).
 
-Each frame dataclass owns its byte layout via ``to_zmq()``/``from_zmq()``. The
-wire-frame dataclasses are pure data plus codecs and import no torch, so the
-protocol contract stays importable and unit-testable on CPU.
+Wire-frame dataclasses live in :mod:`tokenspeed.runtime.pd.epd.entities`: they
+are pure data plus codecs and import no torch, so the protocol contract stays
+importable and unit-testable on CPU.
 """
 
 from __future__ import annotations
 
 import concurrent.futures
-import dataclasses
 import os
-import struct
 import threading
 import time
 from typing import TYPE_CHECKING
@@ -44,13 +42,20 @@ from typing import TYPE_CHECKING
 import requests
 import zmq
 
-from tokenspeed.runtime.disaggregation.base.poll import TransferPoll
-from tokenspeed.runtime.disaggregation.embedding.conn import (
-    EmbeddingManagerArgs,
-    EmbeddingManagerBase,
-    EmbeddingTransferError,
+from tokenspeed.runtime.pd.base.status import TransferPoll
+from tokenspeed.runtime.pd.epd.conn import (
+    MooncakeEmbeddingManagerBase,
 )
-from tokenspeed.runtime.disaggregation.utils import DisaggregationMode, FastQueue
+from tokenspeed.runtime.pd.epd.entities import (
+    REGISTER_ROOM_SENTINEL,
+    EmbeddingArgs,
+    EmbeddingArgsRegisterInfo,
+    EmbeddingChunk,
+    EmbeddingManagerArgs,
+    EmbeddingTransferError,
+    EmbeddingTransferInfo,
+)
+from tokenspeed.runtime.pd.utils import DisaggregationMode, FastQueue
 from tokenspeed.runtime.utils import get_colorful_logger
 from tokenspeed.runtime.utils.env import envs
 from tokenspeed.runtime.utils.network import get_free_port, get_local_ip_by_remote
@@ -82,188 +87,8 @@ def _route_get(bootstrap_addr: str, engine_rank: int, target_dp_group: int):
     return None
 
 
-# Sentinel room (on-wire ascii) for the one-shot registration frame; per-request
-# pre-alloc frames carry an integer room.
-REGISTER_ROOM_SENTINEL = "None"
-
-
-def _pack_ptr(ptr: int) -> bytes:
-    """Encode a device virtual address as 8 little-endian bytes (uint64)."""
-    return struct.pack("Q", ptr)
-
-
-def _unpack_ptr(frame: bytes) -> int:
-    return struct.unpack("Q", frame)[0]
-
-
 def _b(value: object) -> bytes:
     return str(value).encode("ascii")
-
-
-def _bool_frame(value: bool) -> bytes:
-    return b"1" if value else b"0"
-
-
-def _parse_bool(frame: bytes) -> bool:
-    return frame == b"1"
-
-
-@dataclasses.dataclass
-class EmbeddingArgs:
-    """Per-rank buffer/engine info for an embedding endpoint.
-
-    On the encode (sender) side the buffers hold the vision tower's output; on
-    the prefill (receiver) side they are the pre-registered receive buffers. A
-    ``deepstack`` buffer is present only for models that emit deepstack
-    embeddings (e.g. Qwen3.5); ``0`` means absent.
-    """
-
-    engine_rank: int
-    gpu_id: int
-    ib_device: str | None
-    embedding_data_ptr: int
-    embedding_data_len: int
-    deepstack_data_ptr: int = 0
-    deepstack_data_len: int = 0
-
-
-@dataclasses.dataclass
-class EmbeddingChunk:
-    """One item's embedding queued for transfer (encode-side, in-process only).
-
-    Not serialized to ZMQ: it lives on the encode manager's transfer queue. The
-    pointers reference the contiguous ``item.encoded`` / ``item.encoded_deepstack``
-    tensors produced by the vision tower.
-    """
-
-    room: int
-    src_embedding_ptr: int
-    n_tokens: int
-    hidden: int
-    dtype: str
-    nbytes: int
-    src_deepstack_ptr: int = 0
-    deepstack_width: int = 0  # hidden * num_deepstack; 0 == no deepstack
-    deepstack_nbytes: int = 0
-    # CUDA event recorded on the encode loop's stream after the device copy that
-    # filled this chunk's ring slot. The transfer worker waits it before its
-    # one-sided RDMA read so the read never races the copy (ViT->send corruption
-    # hazard). None on CPU/no-CUDA.
-    copy_event: torch.cuda.Event | None = None
-
-
-@dataclasses.dataclass
-class EmbeddingArgsRegisterInfo:
-    """Receiver(prefill)->sender(encode) one-shot endpoint registration.
-
-    Sent once per receiver rank with room == ``REGISTER_ROOM_SENTINEL`` so the
-    encode side records where to PUSH completion status. The buffer pointers
-    are the receiver's base receive buffers (per-item row offsets are carried
-    later, in :class:`EmbeddingTransferInfo`).
-    """
-
-    room: str
-    endpoint: str
-    dst_port: int
-    mooncake_session_id: str
-    dst_embedding_ptr: int
-    dst_deepstack_ptr: int
-
-    def to_zmq(self) -> list[bytes]:
-        return [
-            _b(self.room),
-            _b(self.endpoint),
-            _b(self.dst_port),
-            _b(self.mooncake_session_id),
-            _pack_ptr(self.dst_embedding_ptr),
-            _pack_ptr(self.dst_deepstack_ptr),
-        ]
-
-    @classmethod
-    def from_zmq(cls, msg: list[bytes]) -> "EmbeddingArgsRegisterInfo":
-        return cls(
-            room=msg[0].decode("ascii"),
-            endpoint=msg[1].decode("ascii"),
-            dst_port=int(msg[2].decode("ascii")),
-            mooncake_session_id=msg[3].decode("ascii"),
-            dst_embedding_ptr=_unpack_ptr(msg[4]),
-            dst_deepstack_ptr=_unpack_ptr(msg[5]),
-        )
-
-
-@dataclasses.dataclass
-class EmbeddingTransferInfo:
-    """Receiver(prefill)->sender(encode) per-request pre-allocation frame.
-
-    Tells the encode side where (``dst_embedding_ptr`` already includes this
-    request's row offset) and how big a transfer to issue. The RDMA write is
-    unchecked, so the encode asserts ``n_tokens``/``hidden``/``dtype`` against
-    the tensor it sends, else a mismatch silently writes a truncated/oversized
-    region.
-
-    Row sharding: ``row_start`` + ``n_tokens`` (the SHARD's row count) select
-    the rows the encode writes at the shard-offset ``dst_embedding_ptr``;
-    ``span`` is the image's FULL row count, the cross-side token-count tripwire
-    regardless of shard geometry. Identity frames set ``row_start == 0`` and
-    ``n_tokens == span``. The encode validates shard geometry per frame: a
-    full-span write at a shard-offset pointer would corrupt neighboring image
-    rows in the same buffer.
-    """
-
-    room: int
-    endpoint: str
-    dst_port: int
-    mooncake_session_id: str
-    dst_embedding_ptr: int
-    dst_deepstack_ptr: int
-    n_tokens: int
-    hidden: int
-    dtype: str
-    has_deepstack: bool
-    required_dst_info_num: int
-    row_start: int = 0
-    span: int = 0
-
-    def to_zmq(self) -> list[bytes]:
-        return [
-            _b(self.room),
-            _b(self.endpoint),
-            _b(self.dst_port),
-            _b(self.mooncake_session_id),
-            _pack_ptr(self.dst_embedding_ptr),
-            _pack_ptr(self.dst_deepstack_ptr),
-            _b(self.n_tokens),
-            _b(self.hidden),
-            _b(self.dtype),
-            _bool_frame(self.has_deepstack),
-            _b(self.required_dst_info_num),
-            _b(self.row_start),
-            _b(self.span),
-        ]
-
-    @classmethod
-    def from_zmq(cls, msg: list[bytes]) -> "EmbeddingTransferInfo":
-        # Length-guarded so a truncated frame is logged and dropped by the
-        # bootstrap listener instead of mis-parsing.
-        if len(msg) < 13:
-            raise ValueError(
-                f"malformed EmbeddingTransferInfo frame: {len(msg)} parts < 13"
-            )
-        return cls(
-            room=int(msg[0].decode("ascii")),
-            endpoint=msg[1].decode("ascii"),
-            dst_port=int(msg[2].decode("ascii")),
-            mooncake_session_id=msg[3].decode("ascii"),
-            dst_embedding_ptr=_unpack_ptr(msg[4]),
-            dst_deepstack_ptr=_unpack_ptr(msg[5]),
-            n_tokens=int(msg[6].decode("ascii")),
-            hidden=int(msg[7].decode("ascii")),
-            dtype=msg[8].decode("ascii"),
-            has_deepstack=_parse_bool(msg[9]),
-            required_dst_info_num=int(msg[10].decode("ascii")),
-            row_start=int(msg[11].decode("ascii")),
-            span=int(msg[12].decode("ascii")),
-        )
 
 
 def validate_fanout_frames(
@@ -411,14 +236,14 @@ class MooncakeEmbeddingSender:
         )
 
 
-class MooncakeEmbeddingManagerEncode(EmbeddingManagerBase):
+class MooncakeEmbeddingManagerEncode(MooncakeEmbeddingManagerBase):
     """Encode-side (data source) manager: registers itself to the bootstrap
     server so the prefill receiver can find it, listens for the receiver's
     registration + per-request pre-alloc frames, and a worker thread issues the
     one-sided Mooncake write of the embedding.
     """
 
-    def __init__(self, args: EmbeddingManagerArgs, embedding_args: "EmbeddingArgs"):
+    def __init__(self, args: EmbeddingManagerArgs, embedding_args: EmbeddingArgs):
         super().__init__(args, embedding_args, DisaggregationMode.ENCODE)
         # room -> {receiver_session -> EmbeddingTransferInfo}
         self.transfer_infos: dict[int, dict[str, EmbeddingTransferInfo]] = {}
@@ -709,13 +534,13 @@ class MooncakeEmbeddingManagerEncode(EmbeddingManagerBase):
         )
 
 
-class MooncakeEmbeddingManagerPrefill(EmbeddingManagerBase):
+class MooncakeEmbeddingManagerPrefill(MooncakeEmbeddingManagerBase):
     """Prefill-side (data sink) manager: holds the discovery caches and a thread
     that consumes the encode side's per-request completion-status frames, marking
     the room Success once all expected responses arrive.
     """
 
-    def __init__(self, args: EmbeddingManagerArgs, embedding_args: "EmbeddingArgs"):
+    def __init__(self, args: EmbeddingManagerArgs, embedding_args: EmbeddingArgs):
         super().__init__(args, embedding_args, DisaggregationMode.PREFILL)
         self.required_response_num: dict[int, int] = {}
         self.response_tracker: dict[int, set] = {}
